added doctorCard, SearchDoctor and nearbyDoctors components
