+++ README.md
@@ -1,4 +1,17 @@
 # Yadokari
 
+### TODO
 docker-compose build api
 
+- fix error speciality name in nearby
+- add depond_on for docker
+
+```
+https://stackoverflow.com/questions/52322800/docker-compose-how-to-wait-for-other-service-to-be-ready
+```
+
+- add frontend nearby
+
+- make collection postman
+
+- onKeyUp todo
