+++ frontend/src/pages/aaa.tsx
@@ -1,8 +1,11 @@
 import React, { useState } from 'react';
-import {Layout} from '../app/layout';
-import { searchDoctor } from '../api/doctor';
+import {Layout} from '../components/layout';
+import { searchDoctorFromApi } from '../api/doctor';
 import { Doctor} from '../interface/doctor';
 import { DoctorCard } from '../components/doctorCard';
+import { DoctorList } from '../components/doctorList';
+import { SearchDoctor } from '../components/searchDoctor';
+import { NearbyDoctor } from '../components/nearbyDoctor';
 
 type IndexProps = {
   doctors: Array<Doctor>
@@ -10,64 +13,22 @@ type IndexProps = {
 
 export default function aaa({ doctors}:IndexProps) {
 
-const [stateSearch, setStateSearch] = useState({
-  searchStarted : false,
-  doctorInput : '',
-  doctorFound : false,
-  message : "bjr",
-  doctor :  {} as Doctor | null,
-});
-
-  const handleChangeInput = async (event: any) => {
-    
-    const inputValue = event.currentTarget.previousElementSibling.value
-
-    /*const resDoctor = await fetch(`http://localhost:3001/doctors/${InputValue}`)
-    console.log("resDoctor status ",resDoctor.status)
-    const doctor = await resDoctor.json()*/
-
-    const res = await searchDoctor(inputValue)
-
-    const updatedSearchState = {
-      ...stateSearch
-    };
-
-    if (res.status == 404) {
-      updatedSearchState.message = "no doctor found"
-      updatedSearchState.doctorFound = false
-      
-    } else if (res.status == 200){
-      updatedSearchState.message = "doctor found"
-      updatedSearchState.doctorFound = true
-      updatedSearchState.doctor = res.doctor
-    } else {
-      updatedSearchState.message = "sorry, come back later"
-      updatedSearchState.doctorFound = false
-    }
-
-    setStateSearch(updatedSearchState);
-  }
-
   return (
     <Layout>
-    <main className="flex min-h-screen flex-col items-center justify-between p-24 bg-red-100">
+    <main className="flex min-h-screen flex-col p-8">
 
-      <div className="flex flex-col">
-
-    <input type="text"/>
-    <button onClick={ handleChangeInput}> Search Doctor</button>
-    <div>{stateSearch.message}</div>
-
-    <div className={`${stateSearch.doctorFound ? '' : 'hidden'}`}>
-    
-      <DoctorCard doctor={stateSearch.doctor!}/>
+    <div className="flex flex-col">
+      <h1> nearbyByDoctor</h1>
+      <p> TODO </p>
     </div>
 
-      {doctors.map((d: Doctor, i: number) => (
-                <DoctorCard key = {i} doctor={d}/>
-            )) }
 
-          </div>
+
+      <SearchDoctor/>
+
+
+      <DoctorList doctors={doctors}/>
+      <NearbyDoctor/>
 
     </main>
     </Layout>
